swd: Added ft2232_interface_swd driver declaration
Added ft2232_interface_swd driver declaration with default transport
set to swd_only. It is based on existing ft2232 and jtag_interface. It
contains ft2232_transfer and ft2232_bitbang functions by default, thus
it will work only with ft2232 devices for now. This will be the first
interface type to work with LibSWD.

Change-Id: Ie06ac0430ed588423f32a904b102983ce224a062
Signed-off-by: Tomek CEDRO <cederom@tlen.pl>